feat!: add profile fields and schema class
